Ouput Admin commands
This patch adds a couple of admin commands on outputhost to inspect the
in-memory cg and ack manager states.

(1) The first command is to list all loaded consumer groups on a
specific outputhost
(2) The second command is to read the cg state on an outputhost given
the cgUUID. The cg state includes all in-memory info about this cg and
also the AckManager for all the extents within this CG. This will be
helpful in finding "how many holes are present in the ack manager since
the last ack level".