Add support for draining client connections
This patch simply adds support to drain connections as well.
* To make sure we do this correctly, we maintain a "WritableExtents"
metric which is slightly different from the "OpenExtents" because
a drained extent is still open but is read only.
* When the number of WritableExtents goes to zero, we should send a
DRAIN command on all the connections.
* The connections will simply wait for a timeout period for the drain to
finish. If things finish successfully, we will unload the pathCache
and close all connections automatically. If not, the timeout period will
hit and we will drop the connection object.
* Add an integration test to test this flow.