Fix integration test
Before we drain an extent we need to make sure we give the clients
a chance to drain their side after sending the DRAIN command.

Once all the connections were given a chance (note: this timeout should
be less than the overall drain timeout), start draining the extent.

Increase the message count to 1000 in the drain integration test. We now
consistently pass the test.